home view model set up

+++ app/src/main/java/md/attendance/sl/data/history/HistoryDao.kt
@@ -12,7 +12,7 @@ import androidx.room.Update
 interface HistoryDao {
 
     @Insert
-    suspend fun insertHistory(user: HistoryEntity): Long
+    suspend fun insertHistory(user: HistoryEntity): Int
 
     @Update
     suspend fun updateHistory(user: HistoryEntity)
